requirements.txt
anthropic==0.53.0
attr==0.3.2
brotli==1.1.0
brotlicffi==1.1.0.0
chardet==5.2.0
ConfigParser==7.2.0
contextlib2==21.6.0
cryptography==45.0.4
Cython==3.1.2
dl==0.1.0
docutils==0.21.2
fastapi==0.115.12
filelock==3.18.0
h2==4.2.0
HTMLParser==0.0.2
importlib_resources==6.5.2
ini2toml==0.15
ipython==9.3.0
ipywidgets==8.1.7
js==1.0 
keyring==25.6.0
mock==5.2.0
mod==0.3.0
openai==1.85.0
pandas==2.3.0
path.py==12.5.0
pdfplumber==0.11.6
Pillow==11.2.1
pkg1==0.0.3
protobuf==6.31.1
pydantic==2.11.5
pyOpenSSL==25.1.0
PyPDF2==3.0.1
pytest==8.4.0
redis==6.2.0
seleniumbase==4.39.3
spacy==3.8.7
Sphinx==8.2.3
thread==2.0.5
tomli_w==1.2.0
trove_classifiers==2025.5.9.12
urllib3_secure_extra==0.1.0
uv==0.7.12
xx==3.3.2
zstandard==0.23.0
fastapi
uvicorn[standard]
robotframework
robotframework-databaselibrary
robotframework-jsonlibrary
pymysql
instructor
dotenv